마진 계산 -> 기간별 원가 수정, 캠페인
