+++ src/styles/margin/MarginResultModal.css
@@ -1,4 +1,5 @@
 /* Modal.css */
+/* Modal.css */
 .modal-overlay {
     position: fixed;
     top: 0;
@@ -9,24 +10,50 @@
     display: flex;
     justify-content: center;
     align-items: center;
+    z-index: 1000;
+    /* 다른 요소들 위에 오도록 z-index 추가 */
 }
 
 .modal-content {
     background: white;
     padding: 20px;
-    border-radius: 5px;
+    border-radius: 10px;
+    /* border-radius 값을 10px로 조정 */
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
     width: 1500px;
-    /* 모달 너비 조정 */
+
+    /* ✅ 핵심 수정 사항 */
+    display: flex;
+    /* Flexbox 레이아웃 적용 */
+    flex-direction: column;
+    /* 아이템을 세로로 쌓음 */
+    max-height: 90vh;
+    /* 모달의 최대 높이를 화면 높이의 90%로 제한 */
 }
 
 .close-button {
     background: none;
     border: none;
-    font-size: 16px;
+    font-size: 20px;
+    /* 크기 살짝 키움 */
     cursor: pointer;
-    float: right;
-    /* 오른쪽 정렬 */
+    position: absolute;
+    /* content 기준 위치 */
+    top: 15px;
+    right: 20px;
+}
+
+.modal-body {
+    flex-grow: 1;
+    /* 남은 공간을 모두 차지하도록 설정 */
+    overflow-y: auto;
+    /* 내용이 길어지면 세로 스크롤바 자동 생성 */
+    margin-top: 16px;
+}
+
+.data-table {
+    width: 100%;
+    border-collapse: collapse;
 }
 
 /* MarginResultModal.css */
@@ -210,4 +237,63 @@
     width: 120px;
     /* 입력 필드의 너비 */
     text-align: center;
+}
+
+.data-table tbody td {
+    border-bottom: none
+}
+
+/* 새로운 캠페인 그룹이 시작되는 행의 스타일 */
+.data-table tbody tr.group-start td {
+    /* 위쪽에 굵은 회색 구분선을 추가합니다. */
+    border-top: 2px dashed #e0e0e0;
+}
+
+/* 첫 번째 행에는 위쪽 테두리가 보이지 않도록 처리 */
+.data-table tbody tr:first-child.group-start td {
+    border-top: none;
+}
+
+/* ✅ 그룹의 중간 행 스타일 (가로 구분선 안보이게) */
+.data-table tbody tr.group-middle td {
+    /* 옵션 간의 얇은 구분선을 제거하여 병합된 것처럼 보이게 함 */
+    border-bottom: none;
+}
+
+
+/* 테이블 헤더(th)에 오른쪽 테두리 추가 */
+.data-table thead th {
+    border-right: 1px solid #e9ecef;
+    /* 세로선 스타일 */
+}
+
+/* 마지막 헤더에는 오른쪽 테두리 제거 */
+.data-table thead th:last-child {
+    border-right: none;
+}
+
+/* '캠페인 명' 열의 스타일 */
+.data-table tbody tr td:nth-child(2) {
+    font-size: 16px;
+    /* 기존보다 글자 크기를 약간 키웁니다 */
+    font-weight: 1000;
+    /* 글자를 살짝 두껍게 만듭니다 */
+    color: #6f00f6;
+    /* 밝은 파란색으로 변경합니다 */
+}
+
+.data-table thead th:nth-child(2),
+.data-table tbody td:nth-child(2) {
+    max-width: 250px;
+    /* 최대 너비 (이 값을 조절해 길이를 제어) */
+    white-space: nowrap;
+    /* 텍스트가 한 줄로 표시되도록 강제 */
+    overflow: hidden;
+    /* 너비를 넘어가는 텍스트 숨기기 */
+    text-overflow: ellipsis;
+    /* 숨겨진 텍스트를 ...으로 표시 */
+
+    /* 가독성을 위해 좌측 정렬과 여백을 추가하는 것을 추천합니다. */
+    text-align: left;
+    padding-left: 10px;
 }
